requirements.txt
imagenet-c
numpy
opencv-python-headless
pandas
Pillow
scikit-image
torch
torchvision
